Merged PR 26: Removing extra param & returning empty string when tracelog is undefined
Removing extra param & returning empty string when tracelog is undefined. Turns out adding that extra parameter gives us different results which cause other issues down the line. Decided to return an empty traceLog as a) it doesn't break the system if we don't have a tracelog b) we can pass on the tracelog we need from the debug_traceTransaction method

Related work items: #174